feat: add user management page with CRUD functionality
