+++ apps/web/src/components/pages/Alerts.tsx
@@ -10,7 +10,7 @@ import { Label } from "@/components/ui/label";
 import { Select, SelectContent, SelectItem, SelectTrigger, SelectValue } from "@/components/ui/select";
 import { Switch } from "@/components/ui/switch";
 import { Tabs, TabsContent, TabsList, TabsTrigger } from "@/components/ui/tabs";
-import { Plus, Send, Edit, Trash2, Mail, MessageSquare, Loader2 } from "lucide-react";
+import { Plus, Send, Edit, Trash2, Mail, MessageSquare, Loader2, Bell } from "lucide-react";
 import { NotificationChannel, AlertRule } from "@/types";
 import { useToast } from "@/hooks/use-toast";
 import { notificationChannelService, alertRuleService } from "@/services/alerts.service";
@@ -310,10 +310,15 @@ const Alerts = () => {
 
   return (
     <div className="space-y-6">
-      <div className="flex justify-between items-center">
-        <div>
-          <h1 className="text-3xl font-bold tracking-tight">Alerts & Notifications</h1>
-          <p className="text-muted-foreground">Configure alert rules and notification channels</p>
+      <div className="flex items-center justify-between">
+        <div className="flex items-center gap-3">
+          <div className="p-2 bg-primary/10 rounded-lg">
+            <Bell className="h-6 w-6 text-primary" />
+          </div>
+          <div>
+            <h1 className="text-3xl font-bold tracking-tight">Alerts & Notifications</h1>
+            <p className="text-muted-foreground">Configure alert rules and notification channels</p>
+          </div>
         </div>
       </div>
