+++ apps/web/src/components/pages/SlaveNodes.tsx
@@ -90,7 +90,7 @@ const SlaveNodes = () => {
     switch (status) {
       case 'online': return 'default';
       case 'offline': return 'destructive';
-      case 'syncing': return 'warning';
+      case 'syncing': return 'secondary';
       default: return 'secondary';
     }
   };
@@ -106,10 +106,15 @@ const SlaveNodes = () => {
 
   return (
     <div className="space-y-6">
-      <div className="flex justify-between items-center">
-        <div>
-          <h1 className="text-3xl font-bold tracking-tight">Slave Nodes</h1>
-          <p className="text-muted-foreground">Manage distributed nginx nodes and configuration sync</p>
+      <div className="flex items-center justify-between">
+        <div className="flex items-center gap-3">
+          <div className="p-2 bg-primary/10 rounded-lg">
+            <Server className="h-6 w-6 text-primary" />
+          </div>
+          <div>
+            <h1 className="text-3xl font-bold tracking-tight">Slave Nodes</h1>
+            <p className="text-muted-foreground">Manage distributed nginx nodes and configuration sync</p>
+          </div>
         </div>
         <Dialog open={isDialogOpen} onOpenChange={setIsDialogOpen}>
           <DialogTrigger asChild>
