+++ apps/web/src/components/pages/index.ts
@@ -1,0 +1,17 @@
+// Export all page components
+export { default as ACL } from './ACL';
+export { default as Account } from './Account';
+export { default as Alerts } from './Alerts';
+export { default as Backup } from './Backup';
+export { default as Dashboard } from './Dashboard';
+export { default as Domains } from './Domains';
+export { default as Login } from './Login';
+export { default as Logs } from './Logs';
+export { default as ModSecurity } from './ModSecurity';
+export { default as NotFound } from './NotFound';
+export { default as Performance } from './Performance';
+export { default as SlaveNodes } from './SlaveNodes';
+export { default as SSL } from './SSL';
+export { SSLStats } from './SSLStats';
+export { SSLTable } from './SSLTable';
+export { default as Users } from './Users';
