+++ apps/web/src/routes/$catchall.tsx
@@ -1,5 +1,5 @@
 import { createFileRoute } from '@tanstack/react-router'
-import NotFound from '@/pages/NotFound'
+import NotFound from '@/components/pages/NotFound'
 
 export const Route = createFileRoute('/$catchall')({
   component: NotFound,
