+++ apps/web/src/routes/_auth/account.tsx
@@ -1,4 +1,4 @@
-import Account from '@/pages/Account'
+import { Account } from '@/components/pages'
 import { createFileRoute } from '@tanstack/react-router'
 
 export const Route = createFileRoute('/_auth/account')({
