+++ apps/web/src/routes/_auth/acl.tsx
@@ -1,4 +1,4 @@
-import ACL from '@/pages/ACL'
+import { ACL } from '@/components/pages'
 import { createFileRoute } from '@tanstack/react-router'
 
 export const Route = createFileRoute('/_auth/acl')({
