+++ apps/web/src/routes/_auth/alerts.tsx
@@ -1,4 +1,4 @@
-import Alerts from '@/pages/Alerts'
+import { Alerts } from '@/components/pages'
 import { createFileRoute } from '@tanstack/react-router'
 
 export const Route = createFileRoute('/_auth/alerts')({
