+++ apps/web/src/routes/_auth/nodes.tsx
@@ -1,4 +1,4 @@
-import SlaveNodes from '@/pages/SlaveNodes'
+import SlaveNodes from '@/components/pages/SlaveNodes'
 import { createFileRoute } from '@tanstack/react-router'
 
 export const Route = createFileRoute('/_auth/nodes')({
