+++ apps/web/src/routes/_auth/ssl.tsx
@@ -1,8 +1,17 @@
-import SSL from '@/pages/SSL'
+import SSL from '@/components/pages/SSL'
 import { createFileRoute } from '@tanstack/react-router'
+import { sslQueryOptions } from '@/queries/ssl.query-options'
 
 export const Route = createFileRoute('/_auth/ssl')({
   component: RouteComponent,
+  loader: async ({ context }) => {
+    const { queryClient } = context
+    
+    // Prefetch SSL data but don't await it (allow it to load in background)
+    queryClient.prefetchQuery(sslQueryOptions.all)
+    
+    return {}
+  },
 })
 
 function RouteComponent() {
