+++ apps/web/src/routes/_auth/users.tsx
@@ -1,4 +1,4 @@
-import Users from '@/pages/Users'
+import Users from '@/components/pages/Users'
 import { createFileRoute } from '@tanstack/react-router'
 
 export const Route = createFileRoute('/_auth/users')({
